pyproject.toml
[tool.poetry]
name = "hackaton"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.2"
djangorestframework = "^3.14.0"
django-rest-knox = "^4.2.0"
django-rest-passwordreset = "^1.3.0"
django-filter = "^23.2"
django-cors-headers = "^4.1.0"
zipp = "^3.15.0"
markdown = "^3.4.3"
six = "^1.16.0"
importlib-metadata = "^6.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
